pyproject.toml
[project]
name = "verse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents",
    "livekit-plugins-deepgram",
    "python-dotenv",
    "confluent-kafka",
    "psycopg2-binary",
]
